Looks fixed. Key was points appearing multiple times in sausage.points
During the flood-fill, points are checked by sausageID before being
added to the stack. However, if a point is already on the stack its
sausageID will still be unchanged, so the point will be added to the
stack another time, and later added to sausage.points multiple times,
which affects the sausagePointsIndex which is used in the find-endpoints
algorithm.