Sausages 0 and 1 are now just normal sausages
Sausage 0 used to contain points that weren't in any sausage,
and Sausage 1 contained unsorted points that would eventually end up in
a sausage (i.e. cl<threshold).

This is silly, now we have bool Point.isInASausage and sausageID starts
as -1 to check if we're unsorted, but there's no matching sausage -1