Added zip-file functionality, but slower than raw input
There's a general xyzclcpcs function which reads from an istream.
This istream can be produced by a zip file or a simple ifstream, we
decide which based on filename extensions.

Also added miniz.c, currently only .zip capability. Suspect that gzip
will be faster as we can pipe the data, and also 'gzip --fast' is *much*
faster than default zip (and comparable compression).